feat: move date formatter to helper and test
